+++ ChangeLog.md
@@ -18,6 +18,7 @@
     - Update to pwix:modal v2.5
     - Make sure methods and publications are prefixed with a full namespace
     - Define new scopesCollection configuration parameter
+    - Rename pwix:accounts-hub dependency to pwix:accounts-core, upgrading it to 2.0.0-rc.0
 
 ### 1.8.0